New design
Prior to this change, mbt code was pretty procedural.
Over time, it became hard to understand the overall structure
of the system and its code.
For example, if we had to replace the libgit2 based git integration
to something else, it would have been a tedious task.

New design introduces a modular structure that is
easier to follow and maintain.

See system.go for more information.