Feature: Build/Describe content of a commit
By default `buld|describe commit` commands considers all modules
found in the commit tree.

In this feature introduces a new flag for those commands called
`--content`.

When executed with `--content` flag, only the modules impacted
by the content of the commit are considered for the operation.

We compare the commit with its parent to detect which modules
have been affected. For merge commits with multiple parents,
we use the first parent (This requires verification before
we consider this feature as production ready).