[Feature] Filter modules during commands based on git tree
This feature brings the filtering capability that exists in
`build|describe local` command to `build|describe commit|branch|head`
commands.

Breaking Change
Previously we did a fuzzy match (subsequence) by default when filtering
modules. However, that did not work for cases where you want
to limit the operation to a particular module and that module
name is a prefix of another module (e.g. searching
for app-a in a set of app-a, app-aa returned both app-a and
app-aa).

In this change, we make the default matching behaviour to a
case insensitive exact match. In addition to that the fuzzy
match can be activated by specifying `--fuzzy (-f)` flag.