Fix: Consider file dependencies when calculating the version
Module version is calculated by combining the hash of module
contents, versions of the dependencies and the hashes of its
file dependencies.

This is in contrast to the hash of module contents and versions
of the dependencies used in previous calculation.

Changing the module version when a file dependency is modified
is as important. For example, if we have a file dependency to
a script that builds your module, we should be able to identify
the difference between two instances of the same module built
with different versions of the build script.

Breaking Change
This is a breaking change to how module version is calculated.
If you have used the module version generated by a previous
version of mbt, we highly recommend that you re-run those
operations for those commits in order to get consistent results.

Also Contains
- All messages are in sentence case
- Simplified `Module` data structure by storing `moduleMetadata`
  association.

Closes #59