Feature: Execute user defined commands in modules
User defined commands are specified in spec files
in `commands` section.

Example:

```
commands:
  foo:
    cmd: echo
    args: hello
```

With this configuration in place, running the following command
will execute the command specified as foo in all relevant modules.

```
mbt run-in head --command foo
```

If a module does not specify the command, it will be skipped.
If the command is specific to a particular OS, the constraints can be
specified using `os` property.

Example:
```
commands:
  foo:
    cmd: echo
    args: hello
    os: [darwin, linux]
```

This sub command has all options available in `build|describe` commands.

Important Changes
- Refactored workspace management (i.e. checking out target commits and
  restoring the original at the end of an operation) and process
  management into separate components (WorkspaceManager and
  ProcessManager respectively).

- An info level messages are emitted when checking out a commit and
  restoring the original state of the repository.

Closes #38