Feature: Template helper to convert a map to a list
`kvplist` is a function to convert a `map[string]interface{}`
to a list. This is useful for scenarios where we need to
output the items in a map in a stable manner. In addition to
that, unlike maps, iterating over a list gives us the position
which can be useful in some scenarios.