[fix] Correctly restore detached head after build or custom command
Problem
When repository is in detached state, calling `.Name()` on
`Reference` returned by `Repository.Head()` returns `HEAD`.

When restoring this reference after the build, we pass this value to
`Repository.SetHead` method which basically peels current `HEAD`
to update the value. Since HEAD file is not updated, the diff between
the two trees appear in index as uncommitted.

Solution Overview
- `Reference` returned by `Checkout` has a new member called
`SymbolicName`.
- `SymbolicName` is set to reference name when repository is not in
detached state. Otherwise it has the default empty string.
- During restore, we rely on `SymbolicName` to determine whether to set
HEAD to a symbolic name or a commit SHA in detached mode.

Closes #122